Abstract types

CommonDataModel.AbstractDataset — Type

AbstractDataset is a collection of multidimensional variables (for example a NetCDF or GRIB file)

A data set ds of a type derived from AbstractDataset should implemented at minimum:

Base.key(ds): return a list of variable names a strings

Base.getindex(ds,varname::String): return an array-like data structure (derived from AbstractVariable) of the variables corresponding to varname. This array-like data structure should follow the CF semantics.

ds.attrib: should be a Dict-like data structure with all global attribute names as keys and the corresponding value

ds.dim: should be a Dict-like data structure with all dimension names as keys and the corresponding length

CommonDataModel.AbstractVariable — Type

AbstractVariable{T,N} is a subclass of AbstractArray{T, N}. A variable v of a type derived from AbstractVariable should implement:

v.attrib: should be a Dict-like data structure with all variable attribute names as keys and the corresponding value

name(v): should be the name of variable within the dataset

dimnames(v): should be a iterable data structure with all dimension names

CommonDataModel.dimnames — Method

CommonDatamodel.dimnames(v::AbstractVariable)

Return an iterable of the dimension names of the variable v.

CommonDataModel.group — Method

CommonDatamodel.group(ds::AbstractDataset,groupname::AbstractString)

Return the sub-group data set with the name groupname.

CommonDataModel.groupname — Method

CommonDatamodel.groupname(ds::AbstractDataset)

Name of the group of the data set ds. For a data set containing only a single group, this will be always the root group "/".

CommonDataModel.groups — Method

CommonDatamodel.groups(ds::AbstractDataset)

All the group names of the data set. For a data set containing only a single group, this will be an empty vector of String.

CommonDataModel.name — Method

CommonDatamodel.name(v::AbstractVariable)

Return the name of the variable v as a string.

CommonDataModel.path — Method

CommonDatamodel.path(ds::AbstractDataset)

File path of the data set ds.

CommonDataModel.set_attribute_color — Method

    CommonDataModel.set_attribute_color(color::Symbol)

Set the attribute color. The default color is cyan.

CommonDataModel.set_section_color — Method

    CommonDataModel.set_section_color(color::Symbol)

Set the section color. The default color is red.

CommonDataModel.set_variable_color — Method

    CommonDataModel.set_variable_color(color::Symbol)

Set the variable color. The default color is green.

CommonDataModel.show_attrib — Method

CommonDatamodel.show_attrib(io,a)

Print a list all attributes (key/values pairs) in a to IO stream io. The IO property :level is used for indentation.

CommonDataModel.show_dim — Method

CommonDatamodel.show_dim(io,dim)

Print a list all dimensions (key/values pairs where key is the dimension names and value the corresponding length) in dim to IO stream io. The IO property :level is used for indentation.

CommonDataModel.unlimited — Method

CommonDatamodel.unlimited(ds::AbstractDataset)

Iterator of strings with the name of the unlimited dimension.