Abstract types

In order to implement a new dataset based CommonDataModel.jl one has to create two types derived from: 1 .AbstractVariable: a variable with named dimension and metadata

AbstractDataset: a collection of variable with named dimension, metadata and sub-groups. The sub-groups are also AbstractDataset.

Overview of methods:

get names	get value	write / set value

Dimensions	dimnames	dim	defDim

Attributes	attribnames	attrib	defAttrib

Variables	varnames	variable	defVar

Groups	groupnames	group	defGroup

For read-only datasets, the methods in last column are not implemented.

CommonDataModel.AbstractDataset — Type

AbstractDataset is a collection of multidimensional variables (for example a NetCDF or GRIB file)

A data set ds of a type derived from AbstractDataset should implemented at minimum:

Base.key(ds): return a list of variable names as strings

variable(ds,varname::String): return an array-like data structure (derived from AbstractVariable) of the variables corresponding to varname. This array-like data structure should follow the CF semantics.

dimnames(ds): should be an iterable with all dimension names in the data set ds

dim(ds,name): dimension value corresponding to name

Optionally a data set can have attributes and groups:

attribnames(ds): should be an iterable with all attribute names

attrib(ds,name): attribute value corresponding to name

groupnames(ds): should be an iterable with all group names

group(ds,name): group corresponding to the name

For a writable-dataset, one should also implement:

defDim: define a dimension

defAttrib: define a attribute

defVar: define a variable

defGroup: define a group

CommonDataModel.AbstractVariable — Type

AbstractVariable{T,N} is a subclass of AbstractArray{T, N}. A variable v of a type derived from AbstractVariable should implement:

name(v): should be the name of variable within the data set

dimnames(v): should be a iterable data structure with all dimension names of the variable v

dataset(v): the parent dataset containing v

Base.size(v): the size of the variable

Base.getindex(v,indices...): get the data of v at the provided indices

Optionally a variable can have attributes:

attribnames(v): should be an iterable with all attribute names

attrib(v,name): attribute value corresponding to name

For a writable-dataset, one should also implement:

defAttrib: define a attribute

Base.setindex(v,data,indices...): set the data in v at the provided indices

Base.size — Method

sz = size(var::CFVariable)

Return a tuple of integers with the size of the variable var.

Note

Note that the size of a variable can change, i.e. for a variable with an unlimited dimension.

CommonDataModel._getattrib — Method

_getattrib(ds,v,parentname,attribname,default)

Get a NetCDF attribute, looking also at the parent variable name (linked via the bounds attribute as following the CF conventions). The default value is returned if the attribute cannot be found.

CommonDataModel.attrib — Method

CommonDatamodel.attrib(ds::Union{AbstractDataset,AbstractVariable},attribname::AbstractString)

Return the length of the attributes attribname in the data set ds.

CommonDataModel.attribnames — Method

CommonDatamodel.attribnames(ds::Union{AbstractDataset,AbstractVariable})

Return an iterable of all attribute names in ds.

CommonDataModel.cfvariable — Method

v = cfvariable(ds::NCDataset,varname::AbstractString; <attrib> = <value>)

Return the NetCDF variable varname in the dataset ds as a NCDataset.CFVariable. The keyword argument <attrib> are the NetCDF attributes (fillvalue, missing_value, scale_factor, add_offset, units and calendar) relevant to the CF conventions. By specifing the value of these attributes, the one can override the value specified in the NetCDF file. If the attribute is set to nothing, then the attribute is not loaded and the corresponding transformation is ignored. This function is similar to ds[varname] with the additional flexibility that some variable attributes can be overridden.

Example:

NCDataset("foo.nc","c") do ds
  defVar(ds,"data",[10., 11., 12., 13.], ("time",), attrib = Dict(
      "add_offset" => 10.,
      "scale_factor" => 0.2))
end

# The stored (packed) valued are [0., 5., 10., 15.]
# since 0.2 .* [0., 5., 10., 15.] .+ 10 is [10., 11., 12., 13.]

ds = NCDataset("foo.nc");

@show ds["data"].var[:]
# returns [0., 5., 10., 15.]

@show cfvariable(ds,"data")[:]
# returns [10., 11., 12., 13.]

# neither add_offset nor scale_factor are applied
@show cfvariable(ds,"data", add_offset = nothing, scale_factor = nothing)[:]
# returns [0, 5, 10, 15]

# add_offset is applied but not scale_factor
@show cfvariable(ds,"data", scale_factor = nothing)[:]
# returns [10, 15, 20, 25]

# 0 is declared as the fill value (add_offset and scale_factor are applied as usual)
@show cfvariable(ds,"data", fillvalue = 0)[:]
# return [missing, 11., 12., 13.]

# Use the time units: days since 2000-01-01
@show cfvariable(ds,"data", units = "days since 2000-01-01")[:]
# returns [DateTime(2000,1,11), DateTime(2000,1,12), DateTime(2000,1,13), DateTime(2000,1,14)]

close(ds)

CommonDataModel.dataset — Method

ds = CommonDataModel.dataset(v::AbstractVariable)

Return the data set ds to which a the variable v belongs to.

CommonDataModel.defAttrib — Method

CommonDatamodel.defAttrib(ds::Union{AbstractDataset,AbstractVariable},name::AbstractString,data)

Create an attribute with the name attrib in the data set or variable ds.

CommonDataModel.defDim — Method

CommonDatamodel.defDim(ds::AbstractDataset,name::AbstractString,len)

Create dimension with the name name in the data set ds with the length len. len can be Inf for unlimited dimensions.

CommonDataModel.defGroup — Method

group = CommonDatamodel.defGroup(ds::AbstractDataset,name::AbstractString)

Create an empty sub-group with the name name in the data set ds. The group is a sub-type of AbstractDataset.

CommonDataModel.dim — Method

CommonDatamodel.dim(ds::AbstractDataset,dimname::AbstractString)

Return the length of the dimension dimname in the data set ds.

CommonDataModel.dimnames — Method

CommonDataModel.dimnames(v::AbstractVariable)

Return an iterable of the dimension names of the variable v.

CommonDataModel.dimnames — Method

CommonDatamodel.dimnames(ds::AbstractDataset)

Return an iterable of all dimension names in ds.

CommonDataModel.dims — Method

CommonDatamodel.dims(ds::Union{AbstractDataset,AbstractVariable})

Return a dict-like of all dimensions and their corresponding length defined in the the data set ds (or variable).

CommonDataModel.group — Method

CommonDatamodel.group(ds::AbstractDataset,groupname::AbstractString)

Return the sub-group data set with the name groupname.

CommonDataModel.groupnames — Method

CommonDatamodel.groupnames(ds::AbstractDataset)

All the subgroup names of the data set ds. For a data set containing only a single group, this will be an empty vector of String.

CommonDataModel.groups — Method

CommonDatamodel.groups(ds::AbstractDataset)

Return all sub-group data as a dict-like object.

CommonDataModel.name — Method

CommonDatamodel.name(ds::AbstractDataset)

Name of the group of the data set ds. For a data set containing only a single group, this will be always the root group "/".

CommonDataModel.name — Method

CommonDataModel.name(v::AbstractVariable)

Return the name of the variable v as a string.

CommonDataModel.path — Method

CommonDatamodel.path(ds::AbstractDataset)

File path of the data set ds.

CommonDataModel.set_attribute_color — Method

    CommonDataModel.set_attribute_color(color::Symbol)

Set the attribute color. The default color is cyan.

CommonDataModel.set_section_color — Method

    CommonDataModel.set_section_color(color::Symbol)

Set the section color. The default color is red.

CommonDataModel.set_variable_color — Method

    CommonDataModel.set_variable_color(color::Symbol)

Set the variable color. The default color is green.

CommonDataModel.show_attrib — Method

CommonDatamodel.show_attrib(io,a)

Print a list all attributes (key/values pairs) in a to IO stream io. The IO property :level is used for indentation.

CommonDataModel.show_dim — Method

CommonDatamodel.show_dim(io,dim)

Print a list all dimensions (key/values pairs where key is the dimension names and value the corresponding length) in dim to IO stream io. The IO property :level is used for indentation.

CommonDataModel.time_factor — Method

" tf = time_factor(v::CFVariable)

The time unit in milliseconds. E.g. seconds would be 1000., days would be 86400000. The result can also be nothing if the variable has no time units.

CommonDataModel.unlimited — Method

CommonDatamodel.unlimited(ds::AbstractDataset)

Iterator of strings with the name of the unlimited dimension.

CommonDataModel.variable — Method

CommonDataModel.variable(ds::AbstractDataset,variablename::AbstractString)

Return the variable with the name variablename from the data set ds.

CommonDataModel.varnames — Method

CommonDataModel.varnames(ds::AbstractDataset)

Return an iterable of all the variable name.